requirements.txt
acdh-spacytei==0.0.3
amqp==2.3.2
billiard==3.5.0.4
bleach==2.1.2
boto==2.49.0
boto3==1.9.101
botocore==1.12.101
bz2file==0.98
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.2
cytoolz==0.9.0.1
de-core-news-sm==2.0.0
decorator==4.2.1
dill==0.2.7.1
Django==2.1.3
django-celery-results==1.0.1
django-crispy-forms==1.7.0
django-extensions==2.1.4
django-tables2==1.21.2
djangorestframework==3.7.7
docutils==0.14
entrypoints==0.2.3
ftfy==4.4.3
gensim==3.7.1
html5lib==1.0.1
idna==2.7
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==6.0.0
jupyter-core==4.4.0
kombu==4.2.1
langid==1.1.6
lxml==4.1.1
MarkupSafe==1.0
mistune==0.8.3
msgpack==0.5.6
msgpack-numpy==0.4.1
msgpack-python==0.5.1
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.2
numpy==1.15.4
pandas==0.23.3
pandocfilters==1.4.2
parso==0.3.1
pathlib==1.0.1
pickleshare==0.7.4
plac==0.9.6
preshed==2.0.1
prometheus-client==0.4.2
prompt-toolkit==1.0.15
Pygments==2.2.0
pyreadline==2.1
python-dateutil==2.7.3
pytz==2018.7
pywinpty==0.5.1
pyzmq==17.1.2
qtconsole==4.4.3
regex==2018.1.10
requests==2.20.1
s3transfer==0.2.0
scikit-learn==0.19.0
scipy==0.19.1
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
smart-open==1.8.0
spacy==2.0.18
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.12.1
toolz==0.9.0
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
typing==3.6.4
ujson==1.35
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
